Add multi-page layout with home and placeholder
